+++ src/main/java/com/example/pet/repository/ReviewRepository.java
@@ -9,5 +9,5 @@ import java.util.List;
 @Repository
 public interface ReviewRepository extends JpaRepository<Review, Integer> {
 
-    List<Review> findByMember_MemberId(Long id);
+    List<Review> findByMember_MemberId(int id);
 }
